Implement all CRUD for categories and also some extra endpoints related to soft deletion; Add todos in README.md

+++ src/main/java/com/pixelpolo/hexagon/infrastructure/utils/LocationUtils.java
@@ -14,7 +14,7 @@ public class LocationUtils {
     @Value("${api.version}")
     private String apiVersion;
 
-    public URI getLocation(Integer id, String path) {
+    public URI getLocation(Long id, String path) {
         return URI.create("/api/" + apiVersion + "/" + path + "/" + id);
     }
